# sql: don't copy maps in `prepStmtNamespace.resetTo` if identical