@RaduBerinde 2024-11-04
*[`pkg/sql/conn_executor.go` line 2061 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/134183#-OApJlfyCWYzHptAL0rp:-OAsUmvE9al3E6Jwmi3f:b1d66eu) ([raw file](https://github.com/cockroachdb/cockroach/blob/e958ec85593851adac9bfc3061216d54123acb29/pkg/sql/conn_executor.go#L2061)):*
<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary><blockquote>

`maps.Copy` isn't doing anything special, so I don't think this will be meaningfully faster. I'll leave as is for now, but we can revisit if we want to optimize this further.
</blockquote></details>

Sure, my thinking was that clear is faster than deleting each element. Note that a simple loop that just deletes elements is recognized by the compiler (so it's probably equivalent to clear), but probably not when the loop is more complicated.


<!-- Sent from Reviewable.io -->